+++ packages/dht/.gitignore
@@ -5,3 +5,4 @@ lerna-debug.log
 *.heapprofile
 *.heapsnapshot
 oldsrc
+dist
