+++ packages/dht/protos/ClosestPeers.proto
@@ -6,3 +6,8 @@ message ClosestPeersRequest {
   string peerId = 1;
   string nonce = 2;
 }
+
+message ClosestPeersResponse {
+  repeated string neighborIds = 1;
+  string nonce = 2;
+}
