--- packages/dht/protos/ClosestPeersResponse.proto
@@ -1,8 +1,0 @@
-syntax = "proto3";
-
-option optimize_for = CODE_SIZE;
-
-message ClosestPeersResponse {
-  repeated string neighborIds = 1;
-  string nonce = 2;
-}
