+++ packages/dht/src/protocol/ClosestPeers.ts
@@ -1,0 +1,55 @@
+// @generated by protobuf-ts 2.4.0 with parameter generate_dependencies,
+// @generated from protobuf file "ClosestPeers.proto" (syntax proto3),// tslint:disable
+import { MessageType } from "@protobuf-ts/runtime"
+/**
+ * @generated from protobuf message ClosestPeersRequest
+ */
+export interface ClosestPeersRequest {
+    /**
+     * @generated from protobuf field: string peerId = 1;
+     */
+    peerId: string
+    /**
+     * @generated from protobuf field: string nonce = 2;
+     */
+    nonce: string
+}
+/**
+ * @generated from protobuf message ClosestPeersResponse
+ */
+export interface ClosestPeersResponse {
+    /**
+     * @generated from protobuf field: repeated string neighborIds = 1;
+     */
+    neighborIds: string[]
+    /**
+     * @generated from protobuf field: string nonce = 2;
+     */
+    nonce: string
+}
+// @generated message type with reflection information, may provide speed optimized methods
+class ClosestPeersRequest$Type extends MessageType<ClosestPeersRequest> {
+    constructor() {
+        super("ClosestPeersRequest", [
+            { no: 1, name: "peerId", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "nonce", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ])
+    }
+}
+/**
+ * @generated MessageType for protobuf message ClosestPeersRequest
+ */
+export const ClosestPeersRequest = new ClosestPeersRequest$Type()
+// @generated message type with reflection information, may provide speed optimized methods
+class ClosestPeersResponse$Type extends MessageType<ClosestPeersResponse> {
+    constructor() {
+        super("ClosestPeersResponse", [
+            { no: 1, name: "neighborIds", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
+            { no: 2, name: "nonce", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
+        ])
+    }
+}
+/**
+ * @generated MessageType for protobuf message ClosestPeersResponse
+ */
+export const ClosestPeersResponse = new ClosestPeersResponse$Type()
