--- packages/dht/src/protocol/ClosestPeersRequest.ts
@@ -1,28 +1,0 @@
-// @generated by protobuf-ts 2.4.0,// @generated from protobuf file "ClosestPeersRequest.proto" (syntax proto3),// tslint:disable
-import { MessageType } from "@protobuf-ts/runtime"
-/**
- * @generated from protobuf message ClosestPeersRequest
- */
-export interface ClosestPeersRequest {
-    /**
-     * @generated from protobuf field: string peerId = 1;
-     */
-    peerId: string
-    /**
-     * @generated from protobuf field: string nonce = 2;
-     */
-    nonce: string
-}
-// @generated message type with reflection information, may provide speed optimized methods
-class ClosestPeersRequest$Type extends MessageType<ClosestPeersRequest> {
-    constructor() {
-        super("ClosestPeersRequest", [
-            { no: 1, name: "peerId", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
-            { no: 2, name: "nonce", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
-        ])
-    }
-}
-/**
- * @generated MessageType for protobuf message ClosestPeersRequest
- */
-export const ClosestPeersRequest = new ClosestPeersRequest$Type()
