--- packages/dht/src/protocol/ClosestPeersResponse.ts
@@ -1,28 +1,0 @@
-// @generated by protobuf-ts 2.4.0,// @generated from protobuf file "ClosestPeersResponse.proto" (syntax proto3),// tslint:disable
-import { MessageType } from "@protobuf-ts/runtime"
-/**
- * @generated from protobuf message ClosestPeersResponse
- */
-export interface ClosestPeersResponse {
-    /**
-     * @generated from protobuf field: repeated string neighborIds = 1;
-     */
-    neighborIds: string[]
-    /**
-     * @generated from protobuf field: string nonce = 2;
-     */
-    nonce: string
-}
-// @generated message type with reflection information, may provide speed optimized methods
-class ClosestPeersResponse$Type extends MessageType<ClosestPeersResponse> {
-    constructor() {
-        super("ClosestPeersResponse", [
-            { no: 1, name: "neighborIds", kind: "scalar", repeat: 2 /*RepeatType.UNPACKED*/, T: 9 /*ScalarType.STRING*/ },
-            { no: 2, name: "nonce", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
-        ])
-    }
-}
-/**
- * @generated MessageType for protobuf message ClosestPeersResponse
- */
-export const ClosestPeersResponse = new ClosestPeersResponse$Type()
