+++ packages/dht/tsconfig.node.json
@@ -5,7 +5,8 @@
         "outDir": "dist"
     },
     "include": [
-        "src/**/*"
+        "src/**/*",
+        "google/**/*"
     ],
     "exclude": [
         "dist"
